backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.48
aiomysql==0.2.0
alembic==1.14.1
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
yfinance==0.2.51
openpyxl==3.1.5
